feat: move composables to commons (batch 15)
Move 12 files from amethyst/ui/ to commons with backwards-compatible
re-exports (typealiases for interfaces/enums, delegate functions for
top-level functions/composables):

Moved to commons:
- ZapFormatter (showAmount, showAmountWithZero, constants)
- ZapFormatterNoDecimals (showAmountInteger, showAmountIntegerWithZero)
- PubKeyFormatter (toShortDisplay, toDisplayHexKey, toHexShortDisplay)
- GlowingCard (AnimatedBorderTextCornerRadius)
- LoadingAnimation (LoadingAnimation composable, DefaultAnimationColors)
- ForwardingPainter (forwardingPainter, ForwardingDrawInfo)
- IMessageField interface
- IZapRaiser interface
- IExpiration interface
- BookmarkType enum
- TimeFormatUtils (formatSecondsToTime)

Original files replaced with thin re-export wrappers to maintain
backwards compatibility for existing imports.
