+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/actions/uploads/TimeFormatUtils.kt
@@ -20,20 +20,7 @@
  */
 package com.vitorpamplona.amethyst.ui.actions.uploads
 
-import java.util.Locale
-
-/**
- * Formats seconds into a human-readable time string (M:SS or MM:SS format).
- *
- * @param seconds The number of seconds to format
- * @return Formatted time string (e.g., "0:05", "1:23", "12:45")
- */
-fun formatSecondsToTime(seconds: Int): String {
-    val minutes = seconds / 60
-    val secs = seconds % 60
-    return if (minutes > 0) {
-        String.format(Locale.getDefault(), "%d:%02d", minutes, secs)
-    } else {
-        String.format(Locale.getDefault(), "0:%02d", secs)
-    }
-}
+// Re-export from commons for backwards compatibility
+fun formatSecondsToTime(seconds: Int): String =
+    com.vitorpamplona.amethyst.commons.ui.actions.uploads
+        .formatSecondsToTime(seconds)
