+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/components/ForwardingPainter.kt
@@ -21,59 +21,21 @@
 package com.vitorpamplona.amethyst.ui.components
 
 import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.graphics.DefaultAlpha
 import androidx.compose.ui.graphics.drawscope.DrawScope
 import androidx.compose.ui.graphics.painter.Painter
 
-/**
- * Create and return a new [Painter] that wraps [painter] with its [alpha], [colorFilter], or [onDraw] overwritten.
- */
+// Re-export from commons for backwards compatibility
+typealias ForwardingDrawInfo = com.vitorpamplona.amethyst.commons.ui.components.ForwardingDrawInfo
+
 fun forwardingPainter(
     painter: Painter,
-    alpha: Float = DefaultAlpha,
+    alpha: Float = androidx.compose.ui.graphics.DefaultAlpha,
     colorFilter: ColorFilter? = null,
-    onDraw: DrawScope.(ForwardingDrawInfo) -> Unit = DefaultOnDraw,
-): Painter = ForwardingPainter(painter, alpha, colorFilter, onDraw)
-
-data class ForwardingDrawInfo(
-    val painter: Painter,
-    val alpha: Float,
-    val colorFilter: ColorFilter?,
-)
-
-private class ForwardingPainter(
-    private val painter: Painter,
-    private var alpha: Float,
-    private var colorFilter: ColorFilter?,
-    private val onDraw: DrawScope.(ForwardingDrawInfo) -> Unit,
-) : Painter() {
-    private var info = newInfo()
-
-    override val intrinsicSize get() = painter.intrinsicSize
-
-    override fun applyAlpha(alpha: Float): Boolean {
-        if (alpha != DefaultAlpha) {
-            this.alpha = alpha
-            this.info = newInfo()
+    onDraw: DrawScope.(ForwardingDrawInfo) -> Unit = { info ->
+        with(info.painter) {
+            draw(size, info.alpha, info.colorFilter)
         }
-        return true
-    }
-
-    override fun applyColorFilter(colorFilter: ColorFilter?): Boolean {
-        if (colorFilter == null) {
-            this.colorFilter = colorFilter
-            this.info = newInfo()
-        }
-        return true
-    }
-
-    override fun DrawScope.onDraw() = onDraw(info)
-
-    private fun newInfo() = ForwardingDrawInfo(painter, alpha, colorFilter)
-}
-
-private val DefaultOnDraw: DrawScope.(ForwardingDrawInfo) -> Unit = { info ->
-    with(info.painter) {
-        draw(size, info.alpha, info.colorFilter)
-    }
-}
+    },
+): Painter =
+    com.vitorpamplona.amethyst.commons.ui.components
+        .forwardingPainter(painter, alpha, colorFilter, onDraw)
