+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/components/GlowingCard.kt
@@ -20,30 +20,14 @@
  */
 package com.vitorpamplona.amethyst.ui.components
 
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.PathEffect
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.StrokeJoin
-import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.TextUnit
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
+// Re-export from commons for backwards compatibility
 @Composable
 fun AnimatedBorderTextCornerRadius(
     text: String,
@@ -51,54 +35,10 @@ fun AnimatedBorderTextCornerRadius(
     color: Color = Color.Unspecified,
     textAlign: TextAlign? = null,
     fontSize: TextUnit = 12.sp,
-) {
-    val infiniteTransition = rememberInfiniteTransition()
-    val animatedFloatRestart =
-        infiniteTransition.animateFloat(
-            initialValue = 0f,
-            targetValue = 100f,
-            animationSpec =
-                infiniteRepeatable(
-                    animation = tween(5000, easing = LinearEasing),
-                    repeatMode = RepeatMode.Restart,
-                ),
-        )
-
-    Text(
-        text = text,
-        fontSize = fontSize,
-        modifier =
-            modifier
-                .drawBehind {
-                    val brush =
-                        Brush.sweepGradient(
-                            colors = listOf(Color.Cyan, Color.Magenta, Color.Yellow),
-                        )
-
-                    drawRoundRect(
-                        brush = brush,
-                        style =
-                            Stroke(
-                                width = 2.dp.toPx(),
-                                cap = StrokeCap.Round,
-                                join = StrokeJoin.Round,
-                                pathEffect = PathEffect.dashPathEffect(floatArrayOf(10f, 10f), animatedFloatRestart.value),
-                            ),
-                        cornerRadius =
-                            androidx.compose.ui.geometry
-                                .CornerRadius(6.dp.toPx()),
-                    )
-                }.padding(3.dp),
-        color = color,
-        textAlign = textAlign,
-    )
-}
-
-// Example usage in a composable function:
-@Composable
-@Preview
-fun ExampleAnimatedBorder() {
-    Column {
-        AnimatedBorderTextCornerRadius(text = "Rounded Corners", Modifier)
-    }
-}
+) = com.vitorpamplona.amethyst.commons.ui.components.AnimatedBorderTextCornerRadius(
+    text = text,
+    modifier = modifier,
+    color = color,
+    textAlign = textAlign,
+    fontSize = fontSize,
+)
