+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/components/LoadingAnimation.kt
@@ -20,41 +20,14 @@
  */
 package com.vitorpamplona.amethyst.ui.components
 
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ProgressIndicatorDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.toImmutableList
 
-val DefaultAnimationColors =
-    listOf(
-        Color(0xFF5851D8),
-        Color(0xFF833AB4),
-        Color(0xFFC13584),
-        Color(0xFFE1306C),
-        Color(0xFFFD1D1D),
-        Color(0xFFF56040),
-        Color(0xFFF77737),
-        Color(0xFFFCAF45),
-        Color(0xFFFFDC80),
-        Color(0xFF5851D8),
-    ).toImmutableList()
+// Re-export from commons for backwards compatibility
+val DefaultAnimationColors = com.vitorpamplona.amethyst.commons.ui.components.DefaultAnimationColors
 
 @Composable
 fun LoadingAnimation(
@@ -62,37 +35,9 @@ fun LoadingAnimation(
     circleWidth: Dp = 4.dp,
     circleColors: ImmutableList<Color> = DefaultAnimationColors,
     animationDuration: Int = 1000,
-) {
-    val infiniteTransition = rememberInfiniteTransition()
-
-    val rotateAnimation by
-        infiniteTransition.animateFloat(
-            initialValue = 0f,
-            targetValue = 360f,
-            animationSpec =
-                infiniteRepeatable(
-                    animation =
-                        tween(
-                            durationMillis = animationDuration,
-                            easing = LinearEasing,
-                        ),
-                ),
-            label = "UploadGalleryUploadingAnimation",
-        )
-
-    CircularProgressIndicator(
-        progress = { 1f },
-        modifier =
-            Modifier
-                .size(size = indicatorSize)
-                .rotate(degrees = rotateAnimation)
-                .border(
-                    width = circleWidth,
-                    brush = Brush.sweepGradient(circleColors),
-                    shape = CircleShape,
-                ),
-        color = MaterialTheme.colorScheme.background,
-        strokeWidth = 1.dp,
-        trackColor = ProgressIndicatorDefaults.circularDeterminateTrackColor,
-    )
-}
+) = com.vitorpamplona.amethyst.commons.ui.components.LoadingAnimation(
+    indicatorSize = indicatorSize,
+    circleWidth = circleWidth,
+    circleColors = circleColors,
+    animationDuration = animationDuration,
+)
